backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
docker==7.0.0
mt5linux==1.0.3
rpyc==6.0.0
pyjwt==2.8.0
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
psutil==5.9.8
sqlalchemy==2.0.25
alembic==1.13.1
cryptography==41.0.7
python-dotenv==1.0.0
slowapi==0.1.9
